zeromq: Timeout needs to be in milliseconds for zmq 3.0, close sockets correctly, cleanup

+++ gr-zeromq/lib/pull_source_impl.h
@@ -39,7 +39,7 @@ namespace gr {
       zmq::socket_t   *d_socket;
 
     public:
-      pull_source_impl(size_t itemsize, size_t vlen, char *address, float timeout);
+      pull_source_impl(size_t itemsize, size_t vlen, char *address, int timeout);
       ~pull_source_impl();
 
       int work(int noutput_items,
