+++ gr-zeromq/lib/rep_sink_impl.h
@@ -40,7 +40,7 @@ namespace gr {
       zmq::socket_t   *d_socket;
 
     public:
-      rep_sink_impl(size_t itemsize, size_t vlen, char *address, float timeout, bool blocking);
+      rep_sink_impl(size_t itemsize, size_t vlen, char *address, int timeout, bool blocking);
       ~rep_sink_impl();
 
       int work(int noutput_items,
